Rename FoundationsIndex component
