--- polaris.shopify.com/src/components/FoundationsIndexPage/index.ts
@@ -1,3 +1,0 @@
-import FoundationsIndexPage from './FoundationsIndexPage';
-
-export default FoundationsIndexPage;
