+++ polaris.shopify.com/src/components/FoundationsPage/index.ts
@@ -1,0 +1,3 @@
+import FoundationsPage from './FoundationsPage';
+
+export default FoundationsPage;
